requirements.txt
alembic==1.8.1
anyio==3.6.1
autopep8==1.7.0
awscli==1.22.97
banana-dev==4.0.0
base58==2.1.1
bech32==1.2.0
boto3==1.21.4
botocore==1.24.42
cachetools==5.2.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.7
click==8.1.3
colorama==0.4.3
cryptography==38.0.4
Deprecated==1.2.13
docutils==0.15.2
ecdsa==0.18.0
ECPy==1.2.5
fastapi==0.88.0
greenlet==1.1.2
h11==0.12.0
httpcore==0.13.7
httpx==0.18.2
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
Mako==1.2.1
MarkupSafe==2.1.1
Pillow==9.3.0
pinatapy-vourhey==0.1.8
psycopg2==2.9.5
pyasn1==0.4.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.2
PyJWT==2.4.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==5.4.1
qrcode==7.3.1
requests==2.26.0
rfc3986==1.5.0
rsa==4.7.2
s3transfer==0.5.2
shortuuid==1.0.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.40
starlette==0.22.0
toml==0.10.2
types-Deprecated==1.2.9
typing_extensions==4.4.0
urllib3==1.26.7
uvicorn==0.20.0
websocket-client==1.2.3
websockets==10.3
Werkzeug==2.2.2
wrapt==1.14.1
xrpl-py==1.7.0
xumm-sdk-py==1.0.0
zipp==3.8.1